requirements.txt
better-profanity==0.7.0
argon2-cffi==21.3.0